Wire cart display and order controls to global state (#6)
* Move "Add to Cart" button out of MobileCartControls

* Integrate MobileCartControls with global state

* Create HeaderCartButton component

It's like a regular cart button, but with a cute little badge showing how many items are in the cart! ❤

* Move header into its own component

As part of this, the cart button stopped being its own component and was inserted back into the header markup. This is because there is more noise and boilerplate for a discrete component to work as a trigger for radix-ui's dialog component.

* Wire MobileCartDisplay to show cartContents in state

* Insert addFallSneakersToCart action

There is only one product on sale in this page and the quantity to order is already stored in state. There's no need for logic to be done outside the action.

* Wire Add to Cart button with state action

* Enable clearCart action in MobileCartDialog

* Move MobileCartDialog into own file

* Fix tests by mocking `clearCart` prop

* Update tests to interact with the clear cart button

* Reorganize App.test.tsx into sections

* Add additional tests for App.test.tsx